requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-admin-interface==0.10.5
django-ckeditor==5.7.1
django-colorfield==0.1.15
django-debug-toolbar==1.11
django-extensions==2.1.7
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-modeltranslation==0.13.1
django-rosetta==0.9.2
idna==2.8
Pillow==6.0.0
polib==1.1.0
pydotplus==2.0.2
PyJWT==1.7.1
PySocks==1.7.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
